+++ src/world/block/BlockSelection.cpp
@@ -103,7 +103,7 @@ BlockSelectionBox getTorchBox(const StateID stateId) {
         buildWallTorchTransform(facingValue));
 }
 
-glm::vec3 rotateModelPointX90(const glm::vec3& point, const uint8_t rotation) {
+glm::vec3 rotateModelPointX90(const glm::vec3& point, const uint16_t rotation) {
     switch ((rotation / 90u) % 4u) {
         case 1: return {point.x, 1.0f - point.z, point.y};
         case 2: return {point.x, 1.0f - point.y, 1.0f - point.z};
@@ -113,7 +113,7 @@ glm::vec3 rotateModelPointX90(const glm::vec3& point, const uint8_t rotation) {
     }
 }
 
-glm::vec3 rotateModelPointY90(const glm::vec3& point, const uint8_t rotation) {
+glm::vec3 rotateModelPointY90(const glm::vec3& point, const uint16_t rotation) {
     switch ((rotation / 90u) % 4u) {
         case 1: return {1.0f - point.z, point.y, point.x};
         case 2: return {1.0f - point.x, point.y, 1.0f - point.z};
@@ -123,7 +123,7 @@ glm::vec3 rotateModelPointY90(const glm::vec3& point, const uint8_t rotation) {
     }
 }
 
-glm::vec3 rotateModelPointZ90(const glm::vec3& point, const uint8_t rotation) {
+glm::vec3 rotateModelPointZ90(const glm::vec3& point, const uint16_t rotation) {
     switch ((rotation / 90u) % 4u) {
         case 1: return {1.0f - point.y, point.x, point.z};
         case 2: return {1.0f - point.x, 1.0f - point.y, point.z};
